marcxml.py update serializer for passing tests
dump produces a dict object. In python 3.7 no .data for a dict object is available.
Changed .data to .items which produces a dict_items object consisting of key and value of each schema entry.
The dict_items object must be converted to a dict because dojson module iterates through the items.